pyproject.toml
[tool.poetry]
name = "lido-dao-scripts"
version = "0.1.0"
description = "Repository for Lido DAO Aragon voting scripts."
authors = ["Lido devteam"]
license = "MIT"
readme = "README.md"
homepage = "https://mainnet.lido.fi"

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
eth-brownie = "~1.19.1"

[tool.poetry.dev-dependencies]
black = "~22.6.0"
pre-commit = "~2.19.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py310']
include = '\.pyi?$'
